+++ src/components/DialogWizard/DialogWizard.stories.js
@@ -158,6 +158,47 @@ export const WithOverlay = () => {
   )
 }
 
+export const WithCustomStyles = () => {
+  const steps = [<div>Step 1</div>, <div>Step 2</div>, <div>Step 3</div>]
+
+  const [isDialogOpen, setDialogOpen] = useState(false)
+
+  const handleOpenDialog = () => setDialogOpen(true)
+  const handleCloseDialog = () => setDialogOpen(false)
+
+  const disableBackdropClick = boolean(
+    'Disable backdrop click to close dialog',
+    false,
+  )
+  const disableEscapeKeyDown = boolean(
+    'Disable escape key press to close dialog',
+    false,
+  )
+
+  return (
+    <div>
+      <Button
+        aria-controls='menu-story-default'
+        aria-haspopup='true'
+        onClick={handleOpenDialog}
+      >
+        Open Wizard
+      </Button>
+      <DialogWizard
+        className={styles.wizardExample}
+        isOpen={isDialogOpen}
+        onClose={handleCloseDialog}
+        disableBackdropClick={disableBackdropClick}
+        disableEscapeKeyDown={disableEscapeKeyDown}
+        headerTitle={"Look at me, I'm custom!"}
+        steps={steps}
+        onNextClick={action('next')}
+        classes={{ header: styles.customHeaderExample }}
+      />
+    </div>
+  )
+}
+
 export const Uncontrolled = () => {
   const steps = [<div>Step 1</div>, <div>Step 2</div>, <div>Step 3</div>]